+++ packages/types-env/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @prosopo/types-env
 
+## 2.10.19
+### Patch Changes
+
+- Updated dependencies [71a5952]
+  - @prosopo/keyring@2.9.62
+
+## 2.10.18
+### Patch Changes
+
+  - @prosopo/types-database@4.11.18
+
 ## 2.10.17
 ### Patch Changes